Add test verifying that PT0 funds not stuck
This verifies that PT0 funds are not stuck in the case of a dead branch
whose party tokens have not had their values evaluated.

It does **not** verify that PT0 funds are not stuck in the case of a
live branch (with or without a value) whose party tokens have not had
their values evaluated.