Add another test verifying that PT0 assets are not stuck
As mentioned in the previous commit message, we did not test the case
where the PT0 payout was dependant on the evaluation of a PT_i payout
that evaluated to a non-zero value and was not in a dead branch.
This adds that check veryfying that PT0 will be able to call the `pay()`
method despite no other PTs having been evaluated.